Remove old evaluation stuff
